+++ src/projectgoldstarscalculators/Icon.java
@@ -4,6 +4,7 @@ import java.awt.Toolkit;
 import java.net.URL;
 public class Icon
 {
+    //Returns the Project GoldStars Calculator S icon as an Image
     public static Image getImage()
     {
         URL url = ProjectGoldStarsCalculatorS.class.getResource("/images/ProjectGoldStarsCalculatorSIcon.png");