+++ src/projectgoldstarscalculators/TemperatureConversions.java
@@ -7,7 +7,7 @@ import javax.swing.JOptionPane;
 import javax.swing.JTextField;
 public class TemperatureConversions
 {
-    public static JTextField tf1, tf2, tf3;
+    public static JTextField celsiusField, fahrenheitField, kelvinField;
     
     public TemperatureConversions()
     {
@@ -16,10 +16,10 @@ public class TemperatureConversions
     
     private void temperatureConversions()
     {
-        ProgramWindow additionFrame = new ProgramWindow("Temperature Conversions");
-        additionFrame.setLayout(new GridLayout(3, 3));
-        additionFrame.setSize(750 * ProjectGoldStarsCalculatorS.multiplier, 375 * ProjectGoldStarsCalculatorS.multiplier);
-        additionFrame.setInstructionsMenuBar("Please fill one of the fields and then press the button to convert to the other units.");
+        ProgramWindow frame = new ProgramWindow("Temperature Conversions");
+        frame.setLayout(new GridLayout(3, 3));
+        frame.setSize(750 * ProjectGoldStarsCalculatorS.multiplier, 375 * ProjectGoldStarsCalculatorS.multiplier);
+        frame.setInstructionsMenuBar("Please fill one of the fields and then press the button to convert to the other units.");
         JLabel number1 = new JLabel("Celsius:");
         number1.setForeground(ProjectGoldStarsCalculatorS.color2);
         number1.setFont(ProjectGoldStarsCalculatorS.bodyText1);
@@ -29,24 +29,44 @@ public class TemperatureConversions
         JLabel number3 = new JLabel("Kelvin:");
         number3.setForeground(ProjectGoldStarsCalculatorS.color2);
         number3.setFont(ProjectGoldStarsCalculatorS.bodyText1);
-        tf1 = new JTextField("0");
-        tf1.setFont(ProjectGoldStarsCalculatorS.bodyText2);
-        tf2 = new JTextField("0");
-        tf2.setFont(ProjectGoldStarsCalculatorS.bodyText2);
-        tf3 = new JTextField("0");
-        tf3.setFont(ProjectGoldStarsCalculatorS.bodyText2);
-        additionFrame.add(number1);
-        additionFrame.add(tf1);
-        additionFrame.add(Components.button2("Convert From Celsius", new ConvertFromCelsiusListener()));
-        additionFrame.add(number2);
-        additionFrame.add(tf2);
-        additionFrame.add(Components.button2("Convert From Fahrenheit", new ConvertFromFahrenheitListener()));
-        additionFrame.add(number3);
-        additionFrame.add(tf3);
-        additionFrame.add(Components.button2("Convert From Kelvin", new ConvertFromKelvinListener()));
-        additionFrame.makeVisible();
+        setupCelsiusField();
+        setupFahrenheitField();
+        setupKelvinField();
+        frame.add(number1);
+        frame.add(celsiusField);
+        frame.add(Components.button2("Convert From Celsius", new ConvertFromCelsiusListener()));
+        frame.add(number2);
+        frame.add(fahrenheitField);
+        frame.add(Components.button2("Convert From Fahrenheit", new ConvertFromFahrenheitListener()));
+        frame.add(number3);
+        frame.add(kelvinField);
+        frame.add(Components.button2("Convert From Kelvin", new ConvertFromKelvinListener()));
+        frame.makeVisible();
     }
     
+    private void setupCelsiusField()
+    {
+        celsiusField = new JTextField("0");
+        celsiusField.setFont(ProjectGoldStarsCalculatorS.bodyText2);
+    }
+    
+    private void setupFahrenheitField()
+    {
+        fahrenheitField = new JTextField("0");
+        fahrenheitField.setFont(ProjectGoldStarsCalculatorS.bodyText2);
+    }
+    
+    private void setupKelvinField()
+    {
+        kelvinField = new JTextField("0");
+        kelvinField.setFont(ProjectGoldStarsCalculatorS.bodyText2);
+    }
+    
+    /*
+    This is executed when the user attempts to convert from Celsius.
+    It attempts to convert the value in the Celsius text field to Fahrenheit and Kelvin,
+    and then updates the Fahrenheit and Kelvin text fields to equal the value in the Celsius text field.
+    */
     private class ConvertFromCelsiusListener implements ActionListener
     {
         public void actionPerformed(ActionEvent e)
@@ -54,7 +74,7 @@ public class TemperatureConversions
             double c = 0.0, f = 0.0, k = 0.0;
             try
             {
-                String temp = tf1.getText();
+                String temp = celsiusField.getText();
                 c = Double.parseDouble(temp);
             }
             catch(Exception e2)
@@ -66,13 +86,18 @@ public class TemperatureConversions
             f /= 5;
             f += 32;
             String temp2 = "" + f;
-            tf2.setText(temp2);
+            fahrenheitField.setText(temp2);
             k = c + 273.15;
             temp2 = "" + k;
-            tf3.setText(temp2);
+            kelvinField.setText(temp2);
         }
     }
     
+    /*
+    This is executed when the user attempts to convert from Fahrenheit.
+    It attempts to convert the value in the Fahrenheit text field to Celsius and Kelvin,
+    and then updates the Celsius and Kelvin text fields to equal the value in the Fahrenheit text field.
+    */
     private class ConvertFromFahrenheitListener implements ActionListener
     {
         public void actionPerformed(ActionEvent e)
@@ -80,7 +105,7 @@ public class TemperatureConversions
             double c = 0.0, f = 0.0, k = 0.0;
             try
             {
-                String temp = tf2.getText();
+                String temp = fahrenheitField.getText();
                 f = Double.parseDouble(temp);
             }
             catch(Exception e2)
@@ -92,13 +117,18 @@ public class TemperatureConversions
             c *= 5;
             c /= 9;
             String temp2 = "" + c;
-            tf1.setText(temp2);
+            celsiusField.setText(temp2);
             k = c + 273.15;
             temp2 = "" + k;
-            tf3.setText(temp2);
+            kelvinField.setText(temp2);
         }
     }
     
+    /*
+    This is executed when the user attempts to convert from Kelvin.
+    It attempts to convert the value in the Kelvin text field to Celsius and Fahrenheit,
+    and then updates the Celsius and Fahrenheit text fields to equal the value in the Kelvin text field.
+    */
     private class ConvertFromKelvinListener implements ActionListener
     {
         public void actionPerformed(ActionEvent e)
@@ -106,7 +136,7 @@ public class TemperatureConversions
             double c = 0.0, f = 0.0, k = 0.0;
             try
             {
-                String temp = tf3.getText();
+                String temp = kelvinField.getText();
                 k = Double.parseDouble(temp);
             }
             catch(Exception e2)
@@ -116,12 +146,12 @@ public class TemperatureConversions
             }
             c = k - 273.15;
             String temp2 = "" + c;
-            tf1.setText(temp2);
+            celsiusField.setText(temp2);
             f = c * 9;
             f /= 5;
             f += 32;
             temp2 = "" + f;
-            tf2.setText(temp2);
+            fahrenheitField.setText(temp2);
         }
     }
 }